requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-llms-openrouter
llama-index-vector-stores-chroma
llama-parse
python-dotenv
fastapi
uvicorn
